backend/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
DateTime==5.4
PyJWT==2.8.0
psycopg2-binary==2.9.9
waitress==2.1.2
